backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
firebase-admin==6.4.0
google-auth==2.25.2
python-dotenv==1.0.0
requests==2.31.0

